remove links to iOS version
As apparently, from all the info I could find on the web,
the GPL license is not really compatible with the terms of the app store,
I finally disabled the distribution of the iOS version from it to avoid any trouble.

So I'm removing the corresponding links from the website.

Also, not directly related: renaming MacOSX to macOS on the main page download button.